dashboard/requirements.txt
numpy==1.20.3
pandas==1.2.5
seaborn==0.11.1
plotly==5.1.0
streamlit==0.84.0
kafka-python
scipy
librosa==0.5.1
sounddevice
wavio